Notebook for Running on Colab
Since I currently don't have access to a HPC or nvidia gpu, using colab I can change the runtime to T4 and run the code on it. This is generated with the help of gpt because I didnt want the code to keep cloning the repo everything i ran it and keep things clean. Also i was running into issues with my original code for running it on colab due to  not specifying -arch=sm_75